requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.25.0
python-dotenv>=1.0.0
chromadb>=1.0.0
httpx>=0.25.0
PyPDF2>=3.0.0
python-docx>=1.0.0
Markdown>=3.5
python-multipart>=0.0.6
pydantic>=2.0.0
onnxruntime>=1.16.0
tokenizers>=0.13.0